+++ affinda/models/_models.py
@@ -686,6 +686,15 @@ class Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesO
     :vartype major_group: str
     :ivar soc_code: The 4 digit code representing the SOC2020 classification for this job title.
     :vartype soc_code: int
+    :ivar minor_group_code: The 4 digit code representing the SOC2020 classification for this job
+     title.
+    :vartype minor_group_code: int
+    :ivar sub_major_group_code: The 4 digit code representing the SOC2020 classification for this
+     job title.
+    :vartype sub_major_group_code: int
+    :ivar major_group_code: The 4 digit code representing the SOC2020 classification for this job
+     title.
+    :vartype major_group_code: int
     """
 
     _attribute_map = {
@@ -694,6 +703,9 @@ class Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesO
         "sub_major_group": {"key": "subMajorGroup", "type": "str"},
         "major_group": {"key": "majorGroup", "type": "str"},
         "soc_code": {"key": "socCode", "type": "int"},
+        "minor_group_code": {"key": "minorGroupCode", "type": "int"},
+        "sub_major_group_code": {"key": "subMajorGroupCode", "type": "int"},
+        "major_group_code": {"key": "majorGroupCode", "type": "int"},
     }
 
     def __init__(self, **kwargs):
@@ -708,6 +720,15 @@ class Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesO
         :paramtype major_group: str
         :keyword soc_code: The 4 digit code representing the SOC2020 classification for this job title.
         :paramtype soc_code: int
+        :keyword minor_group_code: The 4 digit code representing the SOC2020 classification for this
+         job title.
+        :paramtype minor_group_code: int
+        :keyword sub_major_group_code: The 4 digit code representing the SOC2020 classification for
+         this job title.
+        :paramtype sub_major_group_code: int
+        :keyword major_group_code: The 4 digit code representing the SOC2020 classification for this
+         job title.
+        :paramtype major_group_code: int
         """
         super(
             Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesOccupationPropertiesClassification,
@@ -718,6 +739,9 @@ class Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesO
         self.sub_major_group = kwargs.get("sub_major_group", None)
         self.major_group = kwargs.get("major_group", None)
         self.soc_code = kwargs.get("soc_code", None)
+        self.minor_group_code = kwargs.get("minor_group_code", None)
+        self.sub_major_group_code = kwargs.get("sub_major_group_code", None)
+        self.major_group_code = kwargs.get("major_group_code", None)
 
 
 class Components1Vvtu5NSchemasInvoicedataPropertiesPaymentamountpaidAllof1(
@@ -8505,6 +8529,15 @@ class JobTitleParsedClassification(msrest.serialization.Model):
     :vartype sub_major_group: str
     :ivar major_group:
     :vartype major_group: str
+    :ivar minor_group_code: The 4 digit code representing the SOC2020 classification for this job
+     title.
+    :vartype minor_group_code: int
+    :ivar sub_major_group_code: The 4 digit code representing the SOC2020 classification for this
+     job title.
+    :vartype sub_major_group_code: int
+    :ivar major_group_code: The 4 digit code representing the SOC2020 classification for this job
+     title.
+    :vartype major_group_code: int
     """
 
     _attribute_map = {
@@ -8513,6 +8546,9 @@ class JobTitleParsedClassification(msrest.serialization.Model):
         "minor_group": {"key": "minorGroup", "type": "str"},
         "sub_major_group": {"key": "subMajorGroup", "type": "str"},
         "major_group": {"key": "majorGroup", "type": "str"},
+        "minor_group_code": {"key": "minorGroupCode", "type": "int"},
+        "sub_major_group_code": {"key": "subMajorGroupCode", "type": "int"},
+        "major_group_code": {"key": "majorGroupCode", "type": "int"},
     }
 
     def __init__(self, **kwargs):
@@ -8527,6 +8563,15 @@ class JobTitleParsedClassification(msrest.serialization.Model):
         :paramtype sub_major_group: str
         :keyword major_group:
         :paramtype major_group: str
+        :keyword minor_group_code: The 4 digit code representing the SOC2020 classification for this
+         job title.
+        :paramtype minor_group_code: int
+        :keyword sub_major_group_code: The 4 digit code representing the SOC2020 classification for
+         this job title.
+        :paramtype sub_major_group_code: int
+        :keyword major_group_code: The 4 digit code representing the SOC2020 classification for this
+         job title.
+        :paramtype major_group_code: int
         """
         super(JobTitleParsedClassification, self).__init__(**kwargs)
         self.soc_code = kwargs.get("soc_code", None)
@@ -8534,6 +8579,9 @@ class JobTitleParsedClassification(msrest.serialization.Model):
         self.minor_group = kwargs.get("minor_group", None)
         self.sub_major_group = kwargs.get("sub_major_group", None)
         self.major_group = kwargs.get("major_group", None)
+        self.minor_group_code = kwargs.get("minor_group_code", None)
+        self.sub_major_group_code = kwargs.get("sub_major_group_code", None)
+        self.major_group_code = kwargs.get("major_group_code", None)
 
 
 class JobTitleParsedParsed(msrest.serialization.Model):
@@ -10193,6 +10241,8 @@ class Rectangle(msrest.serialization.Model):
 
     All required parameters must be populated in order to send to Azure.
 
+    :ivar page_index:
+    :vartype page_index: int
     :ivar x0: Required.
     :vartype x0: float
     :ivar y0: Required.
@@ -10211,6 +10261,7 @@ class Rectangle(msrest.serialization.Model):
     }
 
     _attribute_map = {
+        "page_index": {"key": "pageIndex", "type": "int"},
         "x0": {"key": "x0", "type": "float"},
         "y0": {"key": "y0", "type": "float"},
         "x1": {"key": "x1", "type": "float"},
@@ -10219,6 +10270,8 @@ class Rectangle(msrest.serialization.Model):
 
     def __init__(self, **kwargs):
         """
+        :keyword page_index:
+        :paramtype page_index: int
         :keyword x0: Required.
         :paramtype x0: float
         :keyword y0: Required.
@@ -10229,6 +10282,7 @@ class Rectangle(msrest.serialization.Model):
         :paramtype y1: float
         """
         super(Rectangle, self).__init__(**kwargs)
+        self.page_index = kwargs.get("page_index", None)
         self.x0 = kwargs["x0"]
         self.y0 = kwargs["y0"]
         self.x1 = kwargs["x1"]